Introduce process output streaming support
This change introduces the first bit of functionality to the create,
which enables streaming of data from a process with full integration
with Rust's futures.
The rationale behind this change is that instead of integrating with a
NATS service natively from Rust, we will make use of a proven
implementation in a different language (nats.io has clients for Go and
Python, among others). This decision was made after a survey of the
existing Rust crates, followed by a closer look at the most promising
one (ratsio), turned out rather disappointing. The believe is that the
nats.io developed functionality is of higher quality.
Also playing into this decision is the fact that Polygon has deprecated
NATS streaming altogether and will "soon" roll out WebSocket support
even for Alpaca users. Hence, investing a lot of development effort into
understanding NATS and the existing crates seemed like a effort that
could be better spent elsewhere.

The streaming functionality introduced with this patch basically mounts
a Stream interface as per the futures crate on top of blocking read
calls to the started process' stdout, made by a thread dedicated only to
this purpose. We furthermore provide functionality for mapping a Tokio
style Decoder on top of that stream. Integration with Tokio like this is
mostly done for convenience (and the set of dependencies to its crates
kept very small): Tokio's concept of a Codec defines a "protocol" that
is exactly what we want as well.
Overall the design space for this functionality is rather large. Also,
there are a few crates that could potentially help, but ultimately were
decided against.
- tokio-stdin is basically very similar to our infrastructure, but very
  simple and fixed to the containing process' stdin, which makes it
  unsuitable.
- tokio-stdin-stdout has the same problem
- perhaps the most promising candidate is tokio-file-unix. However, that
  ties us very much to Tokio, the need for this dependency is not really
  obvious, conceptually.