stream.py: Add -s/--server argument to specify a custom list of servers
In order to be able to exercise certain error paths it can be useful to
specify a custom set of servers to connect to.
To that end, this change introduces the -s/--server option that allows
for that. The option can be supplied multiple times in order to allow
for connection to a pool of servers.