Strongly type subscription information
So far we have assumed that clients supply a string as expected by
Polygon when subscribing to events. Dealing with strings in such a way
is both error prone and not very expressive.
To rectify this problem, this change introduces an enum type
representing a subscription, with variants for the different event
types supported by Polygon.