Introduce concept of a client
While we right now only support subscribing to certain (stock related)
events, the Polygon API actually supports many more use cases. In case
we add them in the future, it is nice to have a single go-to entity that
provides access to all those services, after being created with the
appropriate authorization data.
Similar to other crates, this change introduces a Client struct that
fulfills this job. Right now a client can be created with information
from the environment and then provides a concise interface allowing for
the subscription of events.