Remove Python based infrastructure
Attempting to use the Python based infrastructure using the official
nats package revealed that it is not working well and probably never
will. The asyncio based functionality does not handle errors properly
and seemingly does not allow for doing so. Timeouts are also not honored
as they should be.
Ultimately this path did not work out as planned and so we will go back
to going Rust-only. With this change, however, we remove all the Python
specific infrastructure first, such that we can plug in the new
functionality subsequently.