Add first version of ratsio (i.e., NATS) based streaming
This change introduces a first draft of a fully Rust based streaming
infrastructure. We use the ratsio crate to interface with Polygon's NATS
cluster.
Please note that this change is intended solely for the sake of
preservation of this work and documentation of how to work with ratsio.
By now Polygon has enabled support for WebSockets and that is what we
ultimately strive to use moving forward.