Introduce ApiInfo object
In the future we will need more information in addition to the API key.
In particular, in order to support market data streaming we need to know
the URL to connect to in order to receive updates.
To accommodate for this requirement, this change introduces a new
struct, ApiInfo, which encapsulates all this data. A constructor is
added that retrieves this information from the environment.