Rename env.rs to api_info.rs
This change renames the env.rs file to api_info.rs. The main reason is
that the file is some to the ApiInfo struct, but there is no reason for
it to only be created from the environment. So such an association is
misleading. Also, however, the closely related alpaca crate has a
similar structure and we would like to keep similarities visible to
decrease friction when switching between projects.