Add market data stream base URL to ApiInfo
Once we want to stream market data we need a way to identify which
endpoint to connect to first. This endpoint is not necessarily fixed as
Polygon itself uses a different domain than Polygon as provided through
Alpaca.
For this crate this means that we need a way to pass in the URL to
connect to. This change proposes such a way in the form of an
environment variable that backs a newly added field of the recently
introduced ApiInfo struct.