Create directory-style module for event related functionality
So far we have had a file-style module used for the event related
functionality. However, with the recently introduced handshake
code that was already showing signs of a bad structure.
In an attempt to improve on the status quo, this change introduces a
directory-style module for all things event related.