Extend ApiInfo struct with API base URL
In the not-too-distant future we want to issue requests to the Polygon
API, which provides data such as open and closing prices of assets. In
order to do so we need to know the base URL of the API.
This change extends the ApiInfo struct with this information, reading it
from the environment if desired.