Normalize subscriptions as part of Client::subscribe
With the changes to enable tracing of the crate we include certain
function parameter's values in spans. In the case of the subscription
code in particular there may be two problems with that:
1) because we accept an iterator and use its debug representation
   directly in a span, it is the iterators internals that may be
   included in the span
2) multiple subscriptions or subsumed ones would unnecessarily clutter
   the span representation

This change addresses both problems. In essence, we collect the iterator
into a set. This step directly removes all true duplicates. We then
remove all subsumed subscriptions (e.g., a Stock::All of one variant
subsumes all Stock::Symbol(_) versions of the same variant). Lastly,
because we then work with the resulting `HashSet`, we always have a
pleasant to look at type in the resulting span.