Convert Events type into a typedef
We initially defined the Events type to be a dedicated struct just
containing a vector of the inner events. There is no real benefit doing
so over merely having a typedef, which accomplishes the same job but in
a more straight forward manner. To that end, this change defines the
type as a typedef.