Properly export Event variant types
The Event enum that we export contains variants for the different events
clients can subscribe to, each in turn represented by one of three
distinct type.
Unfortunately, these "inner" types were not exported and so while they
could be worked with, they could not be named in user code. This change
rectifies the problem by exporting them correctly.