Merge stock.rs into stream.rs
In the past we have made a half-hearted attempt at separating out
functionality specific to dealing with stocks (as opposed to foreign
exchange and crypto currencies). However, the streaming code was never
written in a way to truly abstract those details. What we ended up with
was a weird split between some stock related definitions and a
supposedly generic stream module that in reality just is equally stock
specific.
With this change we intend to be a bit more honest in that we merge
all the contents of stock.rs into stream.rs, making it more clear that
all streaming is about streaming of data related to stocks. That could
potentially change in the future (though we currently have no plans
doing so), but that would require some redesign.