Introduce Event::symbol method
All types of events united under the Event enum have one thing in
common: they all contain the symbol of the stock the event corresponds
to.
With this change we introduce a method that exploits this commonality,
providing clients with an easy way to directly peek at the symbol of any
Event.