Fix inner tracing span for Client::issue
Investigation of traces has shown that we tracing of the manually
created span inside the Client::issue method is showing up in places
where it should not. This is a known problem we have fought with for
long, but never gotten to the root of it. One workaround is to use the
instrument procedural macro, but that requires a new function definition
and it would require us to clone the request's method and URI in this
particular case, which is not acceptable.
Repeated investigation has now brought to light the issue: we were using
tracing's info_span for creation of the span, assuming that it is
futures aware. That, however, is not the case. To fix this issue we now
switch to using the future aware version, which unfortunately is quite a
bit more verbose.