Serialize aggregates as time zone aware milliseconds
When we initially added support for serializing aggregates (in addition
to the always present deserialization), we merely wanted some
representation. So we went with formatting the result according to
RFC3339.
This change revises this decision, serializing the data back into the
millisecond format we retrieved it in. The main reason for doing so is
that we should not apply any policy in terms of how to serialize and
match the input as closely as possible.