Remove Aggregate::open_price_today member
We have seen many cases where we fail to parse an aggregate that was
received not too long after market open where the 'op' member,
representing today's open price, was simply not set. While the issue has
been reported upstream, no visible progress has been made working
towards a solution.
Because the open price can be retrieved through different means and it
is not actually an important property of the type, this change just
removes the member from the Aggregate struct.