Introduce Client::new constructor
There are legitimate cases where a user does not want to have the API
information taken from the environment or at least done so only once,
and keeping the information around. So far we have not enabled such use
cases, as our only constructor for Client objects was taking this
information directly from the environment.
With this change we add an additional constructor that accepts an
ApiInfo object and which cannot fail.