Use Message type even for regular streaming logic
With this change we switch over to using the Message type (instead of
Event) to implement the core streaming logic. We still expose Events,
but we only create them from a set of Message objects. This step will
allow us to react to status messages that indicate a disconnect in the
future.