Handle disconnect status messages properly
Instead of actually disconnecting by dropping the connection, when
Polygon wants to signal a disconnect to the user it sends a status
message. Silently ignoring those messages as we did in the past can be
fatal, because it means that we will no longer receive any updates
(which we subscribed to) without ever seeing an error or even no
response to a ping.
With this change we evaluate status messages properly and honor
disconnects as we should: by stopping the stream.