Stream a single event at a time as opposed to a Vec
The way Polygon reported multiple events as a vector has bugged us from
the very beginning. Yet, a vector of events was what we exposed as a
single element to retrieve from the stream.
This change finally reworks this API. With it, we return one event at a
time. This change was made possible only because of the unfold logic
that we had introduced earlier to be able to cope with status messages
indicating a disconnect.