Remove Option from Aggregate response type
In the past, whenever an empty set of aggregates was reported, that was
actually represented as a 'null' value. This value is not a valid JSON
representation for a Vec and so we had to wrap it inside an Option.
Polygon folks have finally adjusted their APIs to return an empty
sequence and so this change removes the Option from the response type.