Don't include subscriptions in subscription span
The span we create as part of the market data subscription path includes
a debug representation of the subscriptions. That means that every event
emitted inside this span will also contain this very verbose
representation of the desired subscriptions. It's not only the verbose
nature of the subscriptions, though. It's also their unbounded number,
as a user can provide an arbitrary list of assets to subscribe to.
With this change we remove the subscriptions from the span's fields
altogether. Ultimately it is very unlikely that there will be multiple
subscriptions in one program, as Polygon only allows a single connection
for each account. We also still can see the subscriptions from the
traces: we emit an event containing all the subscriptions inside this
span, but they are formatted in the way Polygon expects them.