Switch to using types from the http crate
This change marks a first step in the process of decoupling the crate a
bit more from hyper. Specifically, it uses request and status code
types from the http crate directly, instead of the re-exports from
hyper. This step will make it easier to add support for a different
backend for issuing requests in the future.