Fix request_tickers test case
Polygon.io changed the company name they return for the ticker AAPL,
making our test fail, as it assumed a different value. With this change
we fix the test by trying to be a bit more flexible in what we accept,
only checking that the string begins with "Apple", but not caring about
the rest.