Allow for creation of ApiInfo object with directly provided data
So far ApiInfo objects had to always be created with information from
the environment, mostly because that is the way we used it. In an
effort to add support for running in WASM environments, where no
environment variables exist, we need a different way to create such an
object.
To that end, this change introduces a new constructor, ApiInfo::new,
that requires the API key to use to be provided.