Remove events::Trade::conditions and events::Quote::condition members
Polygon has apparently changed their API once more, making the
'conditions' field in streamed trade events optional. It really always
has been optional because it was an array that could just contain zero
elements, but that seems to be a representation that was not desired
from the company's side.
In any case, to circumvent this problem (and future problems with this
field), we just remove it. It is actually of very little use to us,
because translating the conditions would require a mapping that has to
be retrieved dynamically, the bindings for which we have not yet
created. Given how much they honor their API "contract" there is really
no point in assuming a mapping from earlier would still hold. As such,
the value as is just useless to us. As a precaution, and for a similar
reason, we also remove the 'condition' field from the quote event.