Add gitlab-ci.yml configuration file
This change adds a configuration file for the Gitlab CI/CD pipeline to
the repository. This file (directly or indirectly) controls the
environment in which to build and what exactly to build.
With the pipeline in place the change also adds a badge indicating the
status of the build on the project's front page.