Use thiserror crate for generating error types
By now the various error types we use throughout the program have become
fairly complex and their specification/definition is a tedious and
repetitive endeavor. In the future we are likely to add even more
complexity into the mix, as we will introduce conditional branches based
on build time settings.
To tame the beast somewhat, this change switches over to using the
thiserror crate for defining the error types. It allows for a fairly
declarative syntax that removes a lot of the repetition and keeps
related properties at the same place.