Introduce dedicated ResponseError type
Having a crate global error type enum encompassing all possible
underlying error in dedicated variants can quickly turn an API into a
hard to understand mess. In our case, having Response::into_result
return such an error type with dozens of variants in it is just bound to
cause confusion, because it is obvious that certain errors (such as
websocket related errors) can't logically be encountered on any of the
endpoints that return a Response object.
With this change we embark on the journey of having more fine grained
error types. Specifically, we introduce a new error type, ResponseError,
that is used solely in the Response::into_result conversion. This scheme
could potentially have other down sides such as too many available
errors and a combinatorial explosion of error types based on the various
possible variants in downstream crates, but we will have to see about
that. For now, it address the aforementioned confusion.