Move hyper specific client logic into separate sub module
In the future we would like to support another client backend:
WebAssembly. This means that the client code will need to be made a bit
more modular.
As a first step towards this goal, this change introduces a new sub
module into client.rs that encapsulates all the hyper specific bits and
pieces.