Add support for compiling for wasm32 target
This change marks another step towards full WASM support: it makes the
create compilable on the wasm32 target (more specifically:
wasm32-unknown-unknown). To do so we have to conditionalize the usage of
all hyper and tungstenite related functionality. We also stub out the
actual WASM backend code. Lastly, we cannot easily run our tests even
with full WASM support because there currently is no way for us to
access environment variables which are needed to make the API key known.
This can be done conceptually with WASI (available through the
wasm32-wasi target), but there we would likely not have browser APIs
available.