Return inner endpoint error as the source
The RequestError type's Endpoint variant should really return the inner
error as its source, otherwise the error itself will not be very useful.
With this change we do that. In order to go about that we require that
the Endpoint trait ensures that the associated Error type implements the
standard Error trait and has a static lifetime, so we bump the required
version to 0.4, which added these constraints.