Introduce web-sys based client backend for WASM support
This change fills in the remaining bits that allow us to issue API
requests from WebAssembly running inside a browser. In particular, we
use web-sys to issue requests using the browser's fetch API. Not that we
do not currently support streaming through a WebSocket connection. This
is more a matter of "not done" as opposed to "impossible". As such, this
support will be added once the need arises.