Reintroduce Option around Aggregate response type
As part of the 0.4.0 release of the crate we removed the Option from the
aggregate response type, as Polygon.io has fixed their API and just
returned an empty vector.
Recently, their "engineers" have gone back to returning null when no
data is available. To that end, this change reintroduces the Option to
deal with these cases.