Fix handling of time stamps captured at daylight savings time
As it turns out the "engineers" over at Polygon.io not only use time
stamps that are based in the US Easter time-zone, they also include
daylight savings time changes in them.
Hence, the previously used fixed-offset based conversion that was
provided by the time-util crate will no longer be sufficient. With this
change we pull in version 0.3 of time-util, which now uses chrono-tz
internally to correctly handle daylight savings time time stamps and
with that fixes this problem.