Simplify date formatting code a bit
As it turns out chrono's DateTime type provides a direct conversion from
Rust's SystemTime. As such, we don't actually need the cumbersome and
roundabout way of converting the latter into the former. With this
change we make use of this conversion to simplify the code. We also
remove a potential error condition that we had to deal with previously.