Introduce proper type representing market status
So far we have treated the market status as an opaque string. The main
motivation for doing so was that Polygon does not specify all possible
variants it can have.
With this change we parse the status properly as either open or closed,
but we leave third variant that will be used for other strings. Doing so
will prevent responses from failing to parse completely, if just one of
the members is faulty.