Fix incorrect time zone being assumed for market time
The way we parse the server time that got returned as part of the market
status response meant that we converted a time stamp that was localized
to New York time to UTC. Such a time zone aware conversion is not
actually what we want, as it will make it impossible to translate times
directly. E.g., 12:00 NY time would be reported as, say, 17:00 UTC, but
it meant that we could no longer make the assumption that 16:00 is when
the market closes regularly.
This change fixes the issue, effectively just dropping the time zone
information.