Transparently handle "delayed" data
On Polygon's free trier data, especially for the current trading day,
may be delayed by 15 minutes. For such data the response received is not
"OK" as we generally expect it to be, but rather "DELAYED".
From our perspective, there is not much of a point in exposing this
information to clients. As such, this change just accepts "DELAYED"
responses as "truly" successful ones.