Introduce error variant for when the rate limit is exceeded
Polygon.io enforces certain rate limits and so when too many requests
are made per unit of time, APIs will return an error. So far we have not
had an explicit variant for such an error, but with this change we add
support for it to each endpoint specific error type.