Fix historic aggregate tests
We no longer have a premium Polygon account and as such we are no longer
able to retrieve data further than two years in the past (or so), as
some of our aggregate tests do.
With this change we adjust these tests to use dates that we can still
access.