Switch to using resolver v2
As it turns out we were implicitly relying on the std/alloc feature of
the chrono dependency to be enabled. When building with the original
v1 resolver this was not a problem, as the feature was enabled by virtue
of our dev-dependency. It specifies the clock feature, which enables
std. However, when used from a Rust edition 2021 project, the build
would fail because this edition uses resolver v2, which has a different
feature resolution process and would compile the dependency without the
std feature (unless compiling tests, which would honor the
dev-dependency specification).
This change adds the std feature to fix this problem. To not trip over
similar issues in the future, we also switch over to using resolver v2
from now on.