Remove serialization support for event and aggregate data
A while back we added support for serializing event and aggregate data.
The main reason for that change was to allow clients to emit back a JSON
representation of the data in question, for one reason or another. At
this point providing this functionality is no longer a priority and the
benefit outweighs the added complexity and maintenance burden. To that
end, this change removes serialization support from the corresponding
types.
If getting a JSON representation becomes important again in the future,
we should devise a way to retrieve the original data, instead of
allowing for reencoding.