Migrate api::aggregates to use Chrono types
This change migrates the api::aggregates module over to using Chrono's
Date<Utc> and DateTime<Utc> instead of a SystemTime. It step marks the
first step in performing this migration for the entire crate. There are
almost no actual benefits to working with a SystemTime and doing so
arguably impacts ergonomics negatively, because so few operations can be
performed on a SystemTime.